requirements.txt
asgiref 
Django 
django-cors-headers 
djangorestframework
djangorestframework_simplejwt 
gunicorn   
packaging 
pillow  
pip   
PyJWT 
sqlparse  
tzdata 
virtualenv
whitenoise 
gunicorn
